requirements.txt
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
Automat==0.7.0
bcrypt==3.1.6
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.2
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-jsonfield==1.1.0
django-mathfilters==0.4.0
django-model-utils==3.1.2
django-notifications-hq==1.4.0
-e git://github.com/v1k45/django-notify-x.git@b4aa03039759126889666a59117e83dcd4cdb374#egg=django_notify_x
django-notify-x2==0.7.0
docutils==0.14
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
jsonfield==2.0.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
more-itertools==5.0.0
mysql-connector==2.2.9
numpy==1.16.1
pandas==0.24.1
Pillow==5.4.1
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.8.1
psycopg2==2.8.2
py==1.7.0
pycparser==2.19
pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
pymongo==3.8.0
pyparsing==2.3.1
pytest==4.2.0
python-dateutil==2.8.0
pytz==2018.7
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
scipy==1.2.1
six==1.12.0
tqdm==4.31.1
twine==1.13.0
urllib3==1.24.2
virtualenv==16.5.0
virtualenv-clone==0.5.3
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.10.11
zope.interface==4.6.0
